Bump vite from 2.6.13 to 2.9.12 in /web
esbuild added support for `riscv64` architecture since 0.14.22 and
vite first used `esbuild@>=0.14.22` since 2.9.0. The latest version of
vite is 2.9.12 as of this commit.

The change regarding to `<script>` is a workaround to vite is trying to
bundle the script despite `/web-config.js` is a external script that
shouldn't be bundled.